+++ simple_merge_sorter.h
@@ -1,12 +1,10 @@
 #pragma once
 
-#include <stddef.h>
-
 #include "sorter.h"
 
 class SimpleMergeSorter : public Sorter {
 public:
-    void sort(std::vector<int32_t>& nums, const size_t processor_numums);
+    void sort(std::vector<int32_t>& nums, const int32_t processor_numums);
 
 private:
     void _merge(const std::vector<int32_t>& left, const std::vector<int32_t>& right, std::vector<int32_t>& dest);
