+++ simple_quick_sorter.h
@@ -3,9 +3,9 @@
 
 class SimpleQuickSorter : public Sorter {
 public:
-    void sort(std::vector<int32_t>& nums, const size_t processor_num);
+    void sort(std::vector<int32_t>& nums, const int32_t processor_num);
 
 private:
-    size_t _partition(std::vector<int32_t>& nums, size_t start, size_t end);
-    void _swap(std::vector<int32_t>& nums, size_t i, size_t j);
+    int32_t _partition(std::vector<int32_t>& nums, int32_t start, int32_t end);
+    void _swap(std::vector<int32_t>& nums, int32_t i, int32_t j);
 };
